ContentModel / Query / Wrangler integration - Major update
* Add ContentWrangler to Engine Servers. Auto listen to dsnode, query data and dispatch events
* Add settings to control content wrangler: content:node_watch, content:model_location, content:use_wrangler
* Dispatch events when dsnode messages come in
* All events return names when called (for verbose logging)
* Add convenience to get engine settings: mEngine.getEngineSettings() for engine.xml settings
* Move work manager to SpriteEngine (since it was always used anyways)
* NodeWatcher can be stopped and started and has some more logging
* Move delayed_node_watcher from essentials to platform

+++ projects/web/cef/src/ds/ui/sprite/web.cpp
@@ -465,7 +465,7 @@ void Web::update(const ds::UpdateParams &p) {
 		ci::gl::Texture::Format fmt;
 		fmt.setMinFilter(GL_LINEAR);
 		fmt.setMagFilter(GL_LINEAR);
-		mPopupTexture = ci::gl::Texture::create(mPopupBuffer, GL_BGRA, mPopupSize.x, mPopupSize.y, fmt);
+		mPopupTexture = ci::gl::Texture::create(mPopupBuffer, GL_BGRA, (int)mPopupSize.x, (int)mPopupSize.y, fmt);
 		mHasPopupBuffer = false;
 		mPopupReady = true;
 	}
